### 2018-08-07 #881 @whedon
[PRE REVIEW]: blendR: Combining a Probability and a Non-Probability Sample in a Capture-Recapture Setting
**Submitting author:** @williamsbenjamin (<a href="http://orcid.org/0000-0001-8474-5066">Benjamin Williams</a>)
**Repository:** <a href="https://github.com/williamsbenjamin/blendR" target ="_blank">https://github.com/williamsbenjamin/blendR</a>
**Version:** 1.0
