### 2018-08-07 96b7f22
fix spelling of "affiliation" in paper.md
