### 2018-08-07 e49462d
spacing edits paper.md
