### 2018-08-07 1b9bd0e
delete Rmd version of paper
