### 2018-08-07 ba2c370
edit affiliation format paper.md
