### 2018-08-07 671a129
more yaml headings in paper.md
