### 2018-08-07 #881 @williamsbenjamin
fixed a quick grammar error.
