### 2018-08-08 #881 @williamsbenjamin
Hello! I believe the article proof is ready to go now. Some reviewers that might be best for this package are: @pdwaggoner or @mdogucu. It may be that @leeper has better thoughts? This package is for combining a non-probability sample and a probability sample in the context of a capture-recapture sampling program. Thanks!
