### 2018-08-08 #881 @leeper
@cimentadaj @pdwaggoner @mdogucu Would any of you be interested in reviewing this for JOSS?
