### 2018-08-08 #881 @pdwaggoner
Thanks for suggesting me as a reviewer, @williamsbenjamin . Sounds interesting. @leeper I would be happy to review this for JOSS, if the need still exists.
