### 2018-08-13 #886 @pdwaggoner
Overall, this looks great. Interesting, simple method. Clearly programmed, and well packaged. I would like to see a couple updates to the paper based on the reviewer guidelines:

1. `Example usage`: the author does a great job of describing functionality and application of the estimators in the context of fisheries. But the author starts the paper with a broad statement of the (clear) value of sampling in statistical inference, despite the application of the methods focusing quite narrowly. I would like to see even a sentence or two on real world application of the estimators in other non-fisheries contexts.

2. `Community guidelines`: though its clear anyone can open an issue ticket in GitHub, it wasn't clearly stated in the paper the willingness and options to do so. Just a little more on how to report bugs, options/outlets for outside contributions and so on would be useful given that this is an open source package.

3. `Statement of need`: I thought the statement of need could be a little clearer and more up front in the manuscript. The value of the method is clear enough, but the need for the package and method got a little lost for me. This point is linked with my comment on `example usage` above. Again, just a sentence or rewording to make the need perfectly clear would be nice.

Other than that, no major issues. Great work @williamsbenjamin ! Over to you @leeper
