### 2018-08-13 #886 @williamsbenjamin
@pdwaggoner thank you very much! These are great comments and I will fix them up right away. Thank you for your time, it is much appreciated!

@leeper should I go ahead and recompile the pdf after I make the changes?
