+++ vignettes/prospectus_bib.bib
@@ -159,9 +159,11 @@
 
 @article{bell_urge_1994,
 	title = {The {Urge} to {Merge}: {Linking} {Vital} {Statistics} {Records} and {Medicaid} {Claims}},
+	volume = {32},
+	number = {10},
 	shorttitle = {The urge to merge},
 	pages = {1004--1018},
-	journal = {Medical care},
+	journal = {Medical Care},
 	author = {Bell, Robert M. and Keesey, Joan and Richards, Toni},
 	year = {1994},
 	file = {3766585.pdf:C\:\\Users\\32443181\\AppData\\Roaming\\Zotero\\Zotero\\Profiles\\dpfiwheo.default\\zotero\\storage\\4KJZ54JV\\3766585.pdf:application/pdf}
