### 2018-08-13 daecbde
updates to paper.md: examples, community usage, statement of need
