### 2018-08-13 f087342
more edits to paper.md
