+++ paper.md
@@ -20,9 +20,9 @@ Statistical sampling plays a vital role in understanding and making inferences w
 
 It is usually easier to obtain a non-probability sample than a probability sample. For example, a frame may not be available, which complicates the selection of a probability sample. A non-probability sample, such as one using data from volunteers, does not require an expensive nonresponse follow-up. The larger the sample, the more information it may contain. This is an attractive option for analysts working with limited resources while studying elusive populations. As a result, statisticians have begun to investigate methods for improving estimation using data from non-probability samples (for example, @elliott_use_2007). One way to improve such estimation is to combine a non-probability sample with a probability sample.
 
-The [blendR](https://github.com/williamsbenjamin/blendR) package (available on [GitHub](https://github.com/williamsbenjamin/blendR)) provides four statistically valid estimators of total when combining a non-probability sample with a probability sample. The estimators are taken from @liu_estimation_2017 and @breidt_model-assisted_2018. The sampling program considers the non-probability sample as a capture sample and the probability sample as a recapture sample, meaning units selected into the non-probability sample can be again sampled into the probability sample. Capture-recapture methodology provides powerful tools to estimate the total number of units in a population [@cren_note_1965]. The goal of the four estimators presented is to make valid estimates of the total of some variable of interest gathered in both samples. The values may disagree for units which are part of both samples (due to measurement error, for example).
+The [blendR](https://github.com/williamsbenjamin/blendR) package (available on [GitHub](https://github.com/williamsbenjamin/blendR)) provides four statistically valid estimators of total when combining a non-probability sample with a probability sample. These estimators have applications in many areas, such as: the internet of things, where a non-probability sample could be taken from devices connected to the internet; insurance claims, where claims could be voluntarily reported; and estimation of the death toll due to a natural disaster, where survivors could self-report deaths in a family. In each of these situations, the estimators from [blendR](https://github.com/williamsbenjamin/blendR) can combine the information from the respective non-probability samples with a probability sample to make more accurate estimates. The prevalence of non-probability samples continues to grow in both academia and industry, due in large part to technological advances and the availability of big data. [blendR](https://github.com/williamsbenjamin/blendR) is needed to allow analysts from a variety of disciplines to use non-probability samples to improve estimation.
 
-These estimators have applications in many areas, such as: the internet of things, where a non-probability sample is taken from devices connected to the internet; insurance claims, where claims can be voluntarily reported; and estimation of the death toll due to a natural disaster, where survivors can self-report deaths in a family. In each of these situations, the estimators from [blendR](https://github.com/williamsbenjamin/blendR) can combine the information from the respective non-probability samples with a probability sample to make more accurate estimates. The prevalence of non-probability samples continues to grow in both academia and industry, due in large part to technological advances and the availability of big data. [blendR](https://github.com/williamsbenjamin/blendR) is needed to allow analysts from a variety of disciplines to improve estimation using non-probability samples.
+The estimators are taken from @liu_estimation_2017 and @breidt_model-assisted_2018. The sampling program considers the non-probability sample as a capture sample and the probability sample as a recapture sample, meaning units selected into the non-probability sample can be again sampled into the probability sample. Capture-recapture methodology provides powerful tools to estimate the total number of units in a population [@cren_note_1965]. The goal of the four estimators presented is to make valid estimates of the total of some variable of interest gathered in both samples. The values may disagree for units which are part of both samples (due to measurement error, for example).
 
[…]
