### 2018-08-13 #886 @williamsbenjamin
@pdwaggoner I have gone through and changed the paper to reflect the comments you had. Please let me know if you think I should go into more details or fix anything else!
