### 2018-08-13 #886 @leeper
@williamsbenjamin The review process is now complete. To finalize your submission and accept your paper in JOSS, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service. To do so:

 3. Post a comment here to @leeper with the DOI for the release.

Let me know if you have any questions about the process.

Oh, and one other really small thing: you'll need to drop the `| file LICENSE` bit of this: https://github.com/williamsbenjamin/blendR/blob/master/DESCRIPTION#L8 and add `LICENSE` to an `.Rbuildignore` file. [`R CMD check` does not allow a LICENSE file for GPL-3](https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#Licensing).
