### 2018-08-13 6b5a660
remove |file LICENSE from DESCRIPTION
