+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 1.0
 Author: Benjamin Williams
 Maintainer: Benjamin Williams <benjamin@smu.edu>
 Description: Provides estimators of total when combining two samples within a capture-recapture framework.
-License: GPL-3 | file LICENSE
+License: GPL-3 
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
