### 2018-08-13 #886 @williamsbenjamin
@leeper Great! I have fixed the `LICENSE` issues. Here is the DOI: 10.5281/zenodo.1344665

Let me know if you need anything else from me!
