+++ README.Rmd
@@ -13,7 +13,7 @@ knitr::opts_chunk$set(
 ```
 
 # blendR
-
+[![DOI](http://joss.theoj.org/papers/10.21105/joss.00886/status.svg)](https://doi.org/10.21105/joss.00886)
 
 The goal of blendR is to provide statistically valid estimators of total (and standard errors) when blending a non-probability sample with a probability sample. The two samples are considered to follow capture-recapture methodology, with the capture sample being the non-probability sample and the recapture sample being the probability sample. This package is based upon research by Liu et al (2017), Breidt, Opsomer, and Huang (2018), and dissertation research by the package author. Additional estimators will be released in future versions.
 
